+++ src/Filters.h
@@ -35,7 +35,7 @@ private:
   uint16_t ascentRateIndex = 0;
   double   altitudeCurr;
   double   altitudeLast;
-  uint64_t ascentRateLast;
+  uint32_t ascentRateLast;
   Eigen::Matrix<double, 2, 1> sensorInputs; // [Ascent_rate , altitude] //z
   Eigen::Matrix<double, 2, 1> currentState; // [Ascent_rate , altitude] //x
   Eigen::Matrix<double, 2, 2> currentCovar; //  P
